Refactor Kubeflow e2e test
PiperOrigin-RevId: 263886080
